+++ R/data.R
@@ -1,0 +1,4 @@
+
+#' Model qualification template.
+#'
+"model_qualification_template"